✨ 내서재에 이미 등록된 도서 중복 추가 방지 구현

+++ src/pages/Modal.js
@@ -67,15 +67,19 @@ const Modal = ({ modal, setModal, data, setData }) => {
   }
 
   const getherData = (key, thumbnail, title, author) => {
-    const newItem = {
-      key: key,
-      thumbnail: thumbnail, 
-      title: title, 
-      author: author,
-      date: date
+    if(data.map(it => it.key.includes(key)).indexOf(true) !== -1) {
+      window.confirm(`내서재에 이미 등록된 도서입니다.`);
+    } else {
+      console.log('새로 등록 가능');
+      const newItem = {
+        key: key,
+        thumbnail: thumbnail, 
+        title: title, 
+        author: author,
+        date: date
+      }
+      setItem(newItem);
     }
-    setItem(newItem);
-    console.log(item);
   }
 
   const addData = () => {
@@ -103,6 +107,7 @@ const Modal = ({ modal, setModal, data, setData }) => {
             onChange={(e) => setKeyword(e.target.value)}
             type="text" 
             placeholder="도서 제목을 입력해주세요."
+            autoComplete="off"
           />
           <button onClick={getData}>검색</button>
           <SearchList result={result} data={data} setData={setData} getherData={getherData}/>
